+++ cloudlets/create_policy_test.go
@@ -7,7 +7,9 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"reflect"
 	"testing"
+	"text/template"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v2/pkg/cloudlets"
 	common "github.com/akamai/cli-common-golang"
@@ -44,6 +46,7 @@ func TestCreatePolicy(t *testing.T) {
 	// This should be removed once a dependency on "github.com/akamai/cli-common-golang" is removed
 	common.App = &cli.App{ErrWriter: ioutil.Discard}
 
+	section := "test_section"
 	pageSize := 1000
 	tests := map[string]struct {
 		init      func(*mockCloudlets, *mockProcessor)
@@ -138,6 +141,7 @@ func TestCreatePolicy(t *testing.T) {
 
 				p.On("ProcessTemplates", TFPolicyData{
 					Name:              "test_policy",
+					Section:           section,
 					CloudletCode:      "ALB",
 					Description:       "version 2 description",
 					GroupID:           234,
@@ -240,6 +244,7 @@ func TestCreatePolicy(t *testing.T) {
 				}, nil).Once()
 				p.On("ProcessTemplates", TFPolicyData{
 					Name:            "test_policy",
+					Section:         section,
 					CloudletCode:    "ER",
 					Description:     "version 2 description",
 					GroupID:         234,
@@ -348,6 +353,7 @@ func TestCreatePolicy(t *testing.T) {
 				}, nil).Once()
 				p.On("ProcessTemplates", TFPolicyData{
 					Name:            "test_policy",
+					Section:         section,
 					CloudletCode:    "CD",
 					Description:     "version 2 description",
 					GroupID:         234,
@@ -427,6 +433,7 @@ func TestCreatePolicy(t *testing.T) {
 				}, nil).Once()
 				p.On("ProcessTemplates", TFPolicyData{
 					Name:              "test_policy",
+					Section:           section,
 					CloudletCode:      "ER",
 					Description:       "version 2 description",
 					GroupID:           234,
@@ -497,6 +504,7 @@ func TestCreatePolicy(t *testing.T) {
 				}, nil).Once()
 				p.On("ProcessTemplates", TFPolicyData{
 					Name:              "test_policy",
+					Section:           section,
 					CloudletCode:      "AP",
 					Description:       "version 2 description",
 					GroupID:           234,
@@ -670,6 +678,7 @@ func TestCreatePolicy(t *testing.T) {
 				}, nil).Once()
 				p.On("ProcessTemplates", TFPolicyData{
 					Name:              "test_policy",
+					Section:           section,
 					CloudletCode:      "ER",
 					Description:       "version 2 description",
 					GroupID:           234,
@@ -695,7 +704,7 @@ func TestCreatePolicy(t *testing.T) {
 			mc := new(mockCloudlets)
 			mp := new(mockProcessor)
 			test.init(mc, mp)
-			err := createPolicy(context.Background(), "test_policy", mc, mp)
+			err := createPolicy(context.Background(), "test_policy", section, mc, mp)
 			if test.withError != nil {
 				assert.True(t, errors.Is(err, test.withError), "expected: %s; got: %s", test.withError, err)
 				return
@@ -716,6 +725,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with ER match rules and activations": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ER",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -804,6 +814,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with ER match rules and single activation": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ER",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -887,6 +898,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ER",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -938,6 +950,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules and invalid escape er": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ER",
 				Description:     `Testing\ exported policy`,
 				GroupID:         12345,
@@ -975,6 +988,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules and invalid escape alb": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ALB",
 				Description:     `Testing\ exported policy`,
 				GroupID:         12345,
@@ -1016,11 +1030,11 @@ func TestProcessPolicyTemplates(t *testing.T) {
 								Continent:       "NA",
 								Country:         "US",
 								Hostname:        "test-hostname",
-								Latitude:        102.78108,
+								Latitude:        tools.Float64Ptr(102.78108),
 								LivenessHosts:   []string{"tf1.test", "tf2.test"},
-								Longitude:       -116.07064,
+								Longitude:       tools.Float64Ptr(-116.07064),
 								OriginID:        "test_origin",
-								Percent:         10,
+								Percent:         tools.Float64Ptr(10),
 								StateOrProvince: tools.StringPtr("MA"),
 							},
 						},
@@ -1045,6 +1059,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules and two alb": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ALB",
 				Description:     `Testing exported policy`,
 				GroupID:         12345,
@@ -1086,11 +1101,11 @@ func TestProcessPolicyTemplates(t *testing.T) {
 								Continent:       "NA",
 								Country:         "US",
 								Hostname:        "test-hostname",
-								Latitude:        102.78108,
+								Latitude:        tools.Float64Ptr(102.78108),
 								LivenessHosts:   []string{"tf1.test", "tf2.test"},
-								Longitude:       -116.07064,
+								Longitude:       tools.Float64Ptr(-116.07064),
 								OriginID:        "test_origin",
-								Percent:         10,
+								Percent:         tools.Float64Ptr(10),
 								StateOrProvince: tools.StringPtr("MA"),
 							},
 						},
@@ -1118,11 +1133,11 @@ func TestProcessPolicyTemplates(t *testing.T) {
 								Continent:       "NA",
 								Country:         "US",
 								Hostname:        "test-hostname",
-								Latitude:        102.78108,
+								Latitude:        tools.Float64Ptr(102.78108),
 								LivenessHosts:   []string{"tf1.test", "tf2.test"},
-								Longitude:       -116.07064,
+								Longitude:       tools.Float64Ptr(-116.07064),
 								OriginID:        "test_origin",
-								Percent:         10,
+								Percent:         tools.Float64Ptr(10),
 								StateOrProvince: tools.StringPtr("MA"),
 							},
 						},
@@ -1147,6 +1162,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy without match rules": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ER",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1158,6 +1174,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules alb": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ALB",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1236,11 +1253,11 @@ func TestProcessPolicyTemplates(t *testing.T) {
 								Continent:       "NA",
 								Country:         "US",
 								Hostname:        "test-hostname",
-								Latitude:        102.78108,
+								Latitude:        tools.Float64Ptr(102.78108),
 								LivenessHosts:   []string{"tf1.test", "tf2.test"},
-								Longitude:       -116.07064,
+								Longitude:       tools.Float64Ptr(-116.07064),
 								OriginID:        "test_origin",
-								Percent:         10,
+								Percent:         tools.Float64Ptr(10),
 								StateOrProvince: tools.StringPtr("MA"),
 							},
 						},
@@ -1265,6 +1282,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules alb and activations": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ALB",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1344,11 +1362,11 @@ func TestProcessPolicyTemplates(t *testing.T) {
 								Continent:       "NA",
 								Country:         "US",
 								Hostname:        "test-hostname",
-								Latitude:        102.78108,
+								Latitude:        tools.Float64Ptr(102.78108),
 								LivenessHosts:   []string{"tf1.test", "tf2.test"},
-								Longitude:       -116.07064,
+								Longitude:       tools.Float64Ptr(-116.07064),
 								OriginID:        "test_origin",
-								Percent:         10,
+								Percent:         tools.Float64Ptr(10),
 								StateOrProvince: tools.StringPtr("MA"),
 							},
 						},
@@ -1389,6 +1407,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy without match rules alb": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "ALB",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1400,6 +1419,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy without match rules fr": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "FR",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1411,6 +1431,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy without match rules CD": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "CD",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1422,6 +1443,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules fr": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "FR",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1487,6 +1509,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules CD": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "CD",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1552,6 +1575,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules vp": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "VP",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1615,6 +1639,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy without match rules vp": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "VP",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1626,6 +1651,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy with match rules ap": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "AP",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1688,6 +1714,7 @@ func TestProcessPolicyTemplates(t *testing.T) {
 		"policy without match rules ap": {
 			givenData: TFPolicyData{
 				Name:            "test_policy_export",
+				Section:         "test_section",
 				CloudletCode:    "AP",
 				Description:     "Testing exported policy",
 				GroupID:         12345,
@@ -1709,6 +1736,9 @@ func TestProcessPolicyTemplates(t *testing.T) {
 					"load-balancer.tmpl": fmt.Sprintf("./testdata/res/%s/load-balancer.tf", test.dir),
 					"variables.tmpl":     fmt.Sprintf("./testdata/res/%s/variables.tf", test.dir),
 					"imports.tmpl":       fmt.Sprintf("./testdata/res/%s/import.sh", test.dir),
+				},
+				AdditionalFuncs: template.FuncMap{
+					"deepequal": reflect.DeepEqual,
 				},
 			}
 			require.NoError(t, processor.ProcessTemplates(test.givenData))