+++ cloudlets/testdata/no_activations_with_two_alb/load-balancer.tf
@@ -76,3 +76,15 @@ resource "akamai_cloudlets_application_load_balancer" "load_balancer_test_origin
   }
 }
 
+resource "akamai_cloudlets_application_load_balancer_activation" "load_balancer_activation_test_origin" {
+  origin_id = akamai_cloudlets_application_load_balancer.load_balancer_test_origin.origin_id
+  network = var.env
+  version = akamai_cloudlets_application_load_balancer.load_balancer_test_origin.version
+}
+
+resource "akamai_cloudlets_application_load_balancer_activation" "load_balancer_activation_test_origin_2" {
+  origin_id = akamai_cloudlets_application_load_balancer.load_balancer_test_origin_2.origin_id
+  network = var.env
+  version = akamai_cloudlets_application_load_balancer.load_balancer_test_origin_2.version
+}
+